add descriptions for options in ChoiceSetting
The long table-like list of options with descriptions is now printed
automagically in the man pages, properly formatted as a list.

The "_" and "help" options are also explained, as well as the default
list of options for a MultiChoiceSetting if there is one.

The default value is also highlighted for ChoiceSettings.